Add a TypeConverterWithParent type, and allow it to be set as DbMap.TypeConverter
This still needs some work, though.  Changing DbMap.TypeConverter to an interface{} type makes it much harder to expose bugs in applications.  I think DbMap should allow a slice of TypeConverter types, but it should maybe keep DbMap.TypeConverter for backward compatibility (maybe deprecate that and remove it in a future version?)